Gate実装

+++ app/Http/Controllers/UserController.php
@@ -22,6 +22,8 @@ class UserController extends Controller
      */
     public function index(UserSearchRequest $request): JsonResponse
     {
+        $this->authorize('viewAny', User::class);
+
         $query = User::query();
         $name = $request->input('name');
         if ($name) {
@@ -61,6 +63,7 @@ class UserController extends Controller
     public function show(int $id): JsonResponse
     {
         $user = User::find($id);
+        $this->authorize('view', $user);
 
         return $user ? response()->json($user) : response()->json(['message' => 'ユーザーが見つかりません'], 404);
     }
@@ -73,6 +76,8 @@ class UserController extends Controller
      */
     public function store(UserRequest $request): JsonResponse
     {
+        $this->authorize('create', User::class);
+
         $data = $request->validated();
         unset($data['password_confirmation']);
 
@@ -94,6 +99,9 @@ class UserController extends Controller
         if (! $user) {
             return response()->json(['message' => 'ユーザーが見つかりません'], 404);
         }
+
+        $this->authorize('update', $user);
+
         $data = $request->validated();
         $user->update($data);
         $updatedUser = $user->fresh();
@@ -109,6 +117,9 @@ class UserController extends Controller
      */
     public function destroy(int $id): Response
     {
+        $user = User::findOrFail($id);
+        $this->authorize('delete', $user);
+
         User::destroy($id);
 
         return response()->noContent();
@@ -150,6 +161,7 @@ class UserController extends Controller
     public function updatePassword(PasswordUpdateRequest $request, int $id): JsonResponse
     {
         $user = User::findOrFail($id);
+        $this->authorize('updatePassword', $user);
 
         if (! empty($request->current_password) && ! Hash::check($request->current_password, $user->password)) {
             return response()->json([
